+++ RELEASE-NOTES.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-* The type-safe object model for secondary indexes has been expanded to incldude multi-indexes.  Calling table.IndexDefineMulti will return an IMultiIndex<TRecord, TIndexType> interface that can be used in multi-index operations, such as GetAll, Between, and EqJoin.
+* The type-safe object model for secondary indexes has been expanded to incldude multi-indexes.  Calling table.IndexDefineMulti will return an IMultiIndex<TRecord, TIndexType> interface that can be used in multi-index operations, such as GetAll, Between, and EqJoin.  [PR #174](https://github.com/mfenniak/rethinkdb-net/issues/174)
 
 
 ## 0.8.0.0 (2014-10-20)